train/requirements.txt
numpy~=1.21.4
torch==1.10.0
torchvision==0.11.1
matplotlib~=3.5.1
tensorflow~=2.7.0
tensorboardX==2.0
terminaltables~=3.1.10
pillow~=8.4.0
tqdm~=4.62.3
